fix compass function(not perfect)

+++ common/common.h
@@ -23,12 +23,12 @@ typedef struct {
 }gsensor_data_t;
 
 typedef struct {
-	u8 mag_xh;
-	u8 mag_xl;
-	u8 mag_yh;
-	u8 mag_yl;
-	u8 mag_zh;
-	u8 mag_zl;
+	s16 mag_xh;
+	s16 mag_xl;
+	s16 mag_yh;
+	s16 mag_yl;
+	s16 mag_zh;
+	s16 mag_zl;
 }mag_data_t;
 
 typedef enum {
